move compatibility feature into its own package
