+++ packages/metadata-builders/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./lookups"
 export * from "./dynamic-builder"
 export * from "./checksum-builder"
-export * from "./compat-builder"
