+++ packages/metadata-compatibility/README.md
@@ -1,0 +1,1 @@
+# @polkadot-api/metadata-compatibility
