+++ packages/metadata-compatibility/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../tsconfig.base",
+  "include": ["src", "tests"],
+  "compilerOptions": {
+    "baseUrl": "src",
+    "resolveJsonModule": true,
+    "paths": {
+      "@/*": ["*"]
+    }
+  }
+}
